requirements.txt
PyYAML==5.3        # CVE-2020-14343 (< 5.4)  ➜ RCE via yaml.load non safe
requests==2.19.1   # GHSA-x84v-xcm2-53pg     ➜ fuite d’Authorization sur redirect
urllib3==1.24.1    # CVE-2019-11324 / 11236  ➜ validation TLS & CRLF redirect